Fix scrolling in browsers with bouncing effect
Browsers with a bounce effect (like Chrome in iOS)
report a negative scrolling value while bouncing
at the top of the page. The same happens when it
bounces down at the bottom (it reports a scroll
value higher than the whole page).

The code detects when the scroll is out of bounds
and prevents it from hiding the menu in that case.